Extract detail screen components to fix SwiftLint file_length and type_body_length
